kaia/requirements.txt
python-telegram-bot>=20.0
anthropic>=0.25.0
groq>=0.5.0
supabase>=2.0.0
apscheduler>=3.10.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
edge-tts>=6.1.0
pydub>=0.25.1
python-dateutil>=2.8.0
